Added logs, minor changes to db seed

+++ services/web/project/service/service.py
@@ -1,16 +1,21 @@
 """
     The module with buisiness logic
 """
-from ..models import *
-from sqlalchemy import desc
 from math import ceil
-from flask import send_from_directory, current_app
 import os
-from re import match, search
+from re import search
+from sqlalchemy import desc
+from ..models import *
+from flask import session, current_app
+from flask_login import current_user
+from werkzeug.security import generate_password_hash, check_password_hash
 
-film_kwargs_pool = ["name", "release_date", "description", "rating", "poster",
-                   "user_id", "genres", "director"]
-director_kwargs_pool = ["first_name", "last_name", "date_of_birth"]
+film_kwargs_pool        = ["name", "release_date", "description", "rating",
+                           "poster", "user_id", "genres", "director"]
+director_kwargs_pool    = ["first_name", "last_name", "date_of_birth"]
+args_list               = ["page", "limit", "sort", "genre", "director",
+                           "date_from", "date_to"]
+
 
 def get_last_id(model):
     return db.session.query(model).order_by(model.id.desc()).first().id
@@ -27,12 +32,19 @@ def get_item_with_id(item, id):
         return None
 
 
+def get_item_with_filter(data, type, filter):
+    query = type.query.filter(filter == data).first()
+    if query is not None:
+        return "item with {} = {} already exists".format(str(filter).split(".")[1], data)
+    return ""
+
 def delete_item_with_id(serialize, type, id):
     item = get_item_with_id(type, id)
     print(item)
     print(serialize.dump(item))
     if item is None:
-        return {"error": "No items with this id"}
+        current_app.logger.warning("No items with id={}".format(id))
+        return {"error": "No items with this id"}, 400
     try:
         db.session.delete(item)
         db.session.commit()
@@ -40,16 +52,22 @@ def delete_item_with_id(serialize, type, id):
         print(item)
         return {"deleted": item}
     except:
-        return {"error": "something went wrong"}
+        current_app.logger.warning("Something went wrong")
+        return {"error": "something went wrong"}, 400
 
 
+def check_author(user):
+    print(user.id, current_user.id, current_user.is_admin)
+    if user.id == current_user.id or current_user.is_admin is True:
+        return True
+    return False
 
 
 def check_director(director_name):
     first_name, last_name = director_name.split(" ")
     if Director.query.where(Director.first_name == first_name).\
             where(Director.last_name == last_name).first() is None:
-        return {"error": "add director first"}
+        return "no director with this name"
     director = Director.query.where(Director.first_name == first_name).\
         where(Director.last_name == last_name).first()
     return director
@@ -63,10 +81,10 @@ def check_film_fileds(**kwargs):
     keys = kwargs.keys()
     if "name" in keys and (len(kwargs["name"])>100 or len(kwargs["name"])<1):
         errors.append("Movie name lenght must be between 1 and 100 characters")
-    if "poster" in keys and (search("([A-Za-z0-9])+(\.jpg|\.png)", kwargs["poster"]) is None):
+    if "poster" in keys and (search(r"([A-Za-z0-9])+(\.jpg|\.png)", kwargs["poster"]) is None):
         errors.append("Not a valid filename for poster")
     if "release_date" in keys:
-        if (search("[1-2][0-9]{3}-[0-1][0-9]-[0-3][0-9]", kwargs["release_date"]) is None):
+        if (search(r"^\d{4}\-(0?[1-9]|1[012])\-(0?[1-9]|[12][0-9]|3[01])$", kwargs["release_date"]) is None):
             errors.append("Valid date format is yyyy-mm-dd")
     if "description" in keys:
         if len(kwargs["description"]) > 1000 or len(kwargs["description"]) < 10:
@@ -87,6 +105,31 @@ def check_film_fileds(**kwargs):
     return errors
 
 
+def check_filter_fileds(**kwargs):
+    errors = []
+    keys = kwargs.keys()
+    for arg in kwargs:
+        if arg == "sort":
+            if kwargs[arg] not in ["date", "r-date", "rating", "r-rating"]:
+                errors.append("wrong sort parameter")
+        if arg == "genre":
+            if kwargs[arg] not in get_genres():
+                errors.append("no such genre")
+        if arg == "director":
+            if check_director(kwargs[arg]) == "no director with this name":
+                errors.append("no director with this name")
+        if arg == "date_from":
+            if (search(r"^\d{4}\-(0?[1-9]|1[012])\-(0?[1-9]|[12][0-9]|3[01])$", kwargs["date_from"]) is None):
+                errors.append("valid date format is yyyy-mm-dd")
+        if arg == "date_to":
+            if (search(r"^\d{4}\-(0?[1-9]|1[012])\-(0?[1-9]|[12][0-9]|3[01])$", kwargs["date_to"]) is None):
+                errors.append("valid date format is yyyy-mm-dd")
+    if "date_from" in kwargs and "date_to" in kwargs:
+        if kwargs["date_from"] > kwargs["date_to"]:
+            errors.append("date_from filter bigger than date_to")
+    return errors
+
+
 def check_director_fileds(**kwargs):
     errors = []
     keys = kwargs.keys()
@@ -98,7 +141,7 @@ def check_director_fileds(**kwargs):
     if "last_name" in keys and (len(kwargs["last_name"])>50 or len(kwargs["last_name"])<1):
         errors.append("Director`s name lenght must be between 1 and 50 characters")
     if "date_of_birth" in keys:
-        if (search("[1-2][0-9]{3}-[0-1][0-9]-[0-3][0-9]", kwargs["date_of_birth"]) is None):
+        if (search(r"[1-2][0-9]{3}-[0-1][0-9]-[0-3][0-9]", kwargs["date_of_birth"]) is None):
             errors.append("Valid date format is yyyy-mm-dd")
     return errors
 
@@ -111,7 +154,6 @@ def paginate_films(query, **kwargs):
         kwargs["limit"] = int(kwargs["limit"])
     else:
         kwargs["limit"] = 10
-
     results = query.count()
     if results > kwargs["limit"]:
         query = query.paginate(page=kwargs["page"], per_page=kwargs["limit"])
@@ -119,9 +161,11 @@ def paginate_films(query, **kwargs):
         kwargs["page"] = 1
         query = query.all()
         if len(query) == 0:
+            current_app.logger.warning("No films requested")
             return 0, kwargs
     kwargs["pages"] = ceil(results / kwargs["limit"])
     return query, kwargs
+    return {"error": "no items with these parameters"}, kwargs
 
 
 def sort_films(base_query, **kwargs):
@@ -134,8 +178,6 @@ def sort_films(base_query, **kwargs):
             query = base_query.query.order_by(Film.rating)
         elif kwargs["sort"] == "rating":
             query = base_query.query.order_by(desc(Film.rating))
-        else:
-            return "Wrong sort parameter"
     else:
         query = base_query.query.order_by(Film.id)
     return query
@@ -151,21 +193,16 @@ def filter_films(query, **kwargs):
         query = query.filter(Film.genres.any(name=kwargs["genre"]))
 
     if "director" in kwargs:
-        try:
-            first_name, last_name = kwargs["director"].split(" ")
-        except:
-            return {"message": "Enter director if format: <first-name> <last-name>"}
+        first_name, last_name = kwargs["director"].split(" ")
         director = db.session.query(Director).where(Director.first_name == first_name,
                                                     Director.last_name == last_name).first()
-        if director is not None:
-            query = query.filter(Film.director_id == director.id)
-        else:
-            return {"message":"Wrong director`s name"}
     return query
 
 
 def get_result(query, **kwargs):
+    current_app.logger.warning(query)
     query = films.dump(list(query))
+    current_app.logger.warning(query)
     for item in query:
         if item["director"] is None:
             item["director"] = "Unknown"
@@ -175,54 +212,68 @@ def get_result(query, **kwargs):
 
 
 def get_films(**kwargs):
-    args_list = ["page", "limit", "sort", "genre", "director", "date_from", "date_to"]
     for item in kwargs:
         if item not in args_list:
-            return {"message":"Wrong parameters"}
+            current_app.logger.warning("Wrong parameters")
+            return {"message":"Wrong parameters"}, 400
+    errors = check_filter_fileds(**kwargs)
+    if len(errors)>0:
+        current_app.logger.warning("errors: {}".format(errors))
+        return errors, 400
     base_query = Film
-
     #sorting
     query = sort_films(base_query, **kwargs)
 
     #filtering
     query = filter_films(query, **kwargs)
+    current_app.logger.warning(query)
 
     #pagination kwargs
     query, kwargs = paginate_films(query, **kwargs)
-
+    if query == 0:
+        return query, 400
     return get_result(query, **kwargs)
 
 
 def get_film(id):
     movie = get_item_with_id(Film, id)
     if movie is None:
-        return {"error": "No items with whis id"}
+        current_app.logger.warning("No films with id={}".format(id))
+        return {"error": "No items with whis id"}, 400
     result = film.dump(movie)
     return result
 
 def get_film_img(id):
     movie = get_item_with_id(Film, id)
-    if os.path.isfile("project/static/" + movie.poster) is False:
-        return None
     return movie.poster
 
 
 def add_film(**kwargs):
+    #validate_fields
     errors = check_film_fileds(**kwargs)
+
     if "description" not in kwargs:
         kwargs["description"] = ""
-    if len(errors) > 0:
-        return {"errors": errors}
     if len(kwargs) != 8:
-        return {"errors": "check parameters again"}
+        errors.append("check parameters again")
+
+    check_name = get_item_with_filter(kwargs["name"], Film, Film.name)
+    if len(check_name)>0:
+        errors.append(check_name)
+
+    #check director
     director = check_director(kwargs["director"])
-    if type(director) == type({}):
-        return director
+    if isinstance(director, str):
+        errors.append(director)
+
+    if len(errors) > 0:
+        current_app.logger.warning("errors: {}".format(errors))
+        return {"errors": errors}, 400
+
     user = get_item_with_id(User, kwargs["user_id"])
     genres = get_genres_list(kwargs["genres"])
     id = get_last_id(Film)
     obj = {
-        "id": id+1,
         "name": kwargs["name"],
         "release_date": kwargs["release_date"],
         "description": kwargs["description"],
@@ -235,23 +286,30 @@ def add_film(**kwargs):
     obj.genres.extend(genres)
     db.session.add(obj)
     db.session.commit()
+    current_app.logger.info("Added the film")
     return {"film": film.dump(obj)}
 
 
 def update_film(id, **kwargs):
     errors = check_film_fileds(**kwargs)
-    if len(errors) > 0:
-        return {"errors": errors}
     keys = kwargs.keys()
     if "director" in keys:
         director = check_director(kwargs["director"])
-        if type(director) == type({}):
-            return director
+        if isinstance(director, str):
+            errors.append(director)
         kwargs["director"] = director
     print(kwargs)
     movie = get_item_with_id(Film, id)
+    check_author(movie.user)
+
+    if check_author(movie.user) is False:
+        errors.append("Not an author or admin")
     if movie is None:
-        return {"error": "No items with this id"}
+        errors.append("No items with this id")
+
+    if len(errors) > 0:
+        current_app.logger.warning("errors: {}".format(errors))
+        return {"errors": errors}, 400
     if "genres" in keys:
         genres = get_genres_list(kwargs["genres"])
         movie.genres = genres
@@ -264,27 +322,44 @@ def update_film(id, **kwargs):
 
 
 def delete_film(id):
+    movie = get_item_with_id(Film, id)
+    if check_author(movie.user) is False:
+        current_app.logger.warning("user with id={} is not an author or admin".format(current_user.id))
+        return {"error": "Not an author or admin"}, 400
     return delete_item_with_id(film, Film, id)
 
 
-def get_Directors(page, limit=10):
+def get_directors(page, limit=10):
     results = Director.query.count()
+    if results < 1:
+        current_app.logger.warning("No directors in db")
+        return {"error": "no directors found"}, 400
     result = Director.query.order_by(Director.id).paginate(page=int(page), per_page=limit)
     return {"pages":ceil(results/limit), "page":page, "directors":directors.dump(result)}
 
 
-def get_Director(id):
-    result = director.dump(get_item_with_id(Director, id))
+def get_director(id):
+    result = get_item_with_id(Director, id)
+    if result is None:
+        current_app.logger.warning("No director with id={}".format(id))
+        return {"error": "no directors with this id"}, 400
+    result = director.dump(result)
     return result
 
 
 def create_director(**kwargs):
     errors = check_director_fileds(**kwargs)
+    error = get_item_with_filter(kwargs["first_name"], Director, Director.first_name)
+    if len(error)>0:
+        errors.append(error)
+    error = get_item_with_filter(kwargs["last_name"], Director, Director.last_name)
+    if len(error)>0:
+        errors.append(error)
     if len(errors)>0:
-        return {"errors": errors}
-    id = get_last_id(Director)
+        current_app.logger.warning("errors: {}".format(errors))
+        return {"errors": errors}, 400
+
     obj = {
-        "id": id+1,
         "first_name": kwargs["first_name"],
         "last_name": kwargs["last_name"],
         "date_of_birth": kwargs["date_of_birth"]
@@ -298,10 +373,12 @@ def create_director(**kwargs):
 def update_director(id, **kwargs):
     director_update = get_item_with_id(Director, id)
     if director_update is None:
-        return {"error": "No items with this id"}
+        current_app.logger.warning("No director with id={}".format(id))
+        return {"error": "No items with this id"}, 400
     errors = check_director_fileds(**kwargs)
     if len(errors) > 0:
-        return {"errors": errors}
+        current_app.logger.warning("errors: {}".format(errors))
+        return {"errors": errors}, 400
 
     keys = kwargs.keys()
     for item in keys:
@@ -315,11 +392,52 @@ def delete_director(id):
 
 
 def get_genres():
+    """
+    Returns list of genres from db
+    @return: genres
+    @rtype: list
+    """
     result = Genre.query.all()
+    if result is None:
+        current_app.logger.warning("No genres in database")
+        return []
     genres = [genre.name for genre in result]
     return genres
 
 
-def get_genres_list(genres):
+def get_genres_list(items):
+    """
+    Returns film genres as list
+
+    @param items:
+    @type string:
+    @return: genres
+    @rtype: list
+    """
     return [Genre.query.where(Genre.name == item.lower()).first()
-            for item in genres.split(",")]
+            for item in items.split(",")]
+
+
+def signup(username, password):
+    with current_app.app_context():
+        new_user = User(username=username)
+        new_user.password_hash = password
+
+        db.session.add(new_user)
+        db.session.commit()
+
+        session['user_id'] = new_user.id
+        return new_user.to_dict()
+
+
+def get_user_data():
+    usr = get_item_with_id(User, current_user.id)
+    if usr is not None:
+        return {"user:":models.user.dump(usr)}
+
+
+def set_user_password(password):
+    current_user.password = generate_password_hash(password)
+    usr = get_item_with_id(User, current_user.id)
+    usr.pasword = generate_password_hash(password)
+    db.session.commit()
